Levenshtein distance calculation moved to a separate file
